==> Notebooks/2_Sample_Data.ipynb <==
# Sample Symptom and Vax data

1. Load Sampled event data that exists in SQL database
2. Load full symptom and vax dataframes
3. Use VAERS_ID from sampled event data to filter symptom and vax dataframes
4. Save Sampled symptoms and vax data

import pandas as pd

# load data
sample_event_data_df = pd.read_csv("../Data/Clean/sample_event_data.csv")
sympt_data_df = pd.read_csv("../Data/Clean/sympt_data.csv")
vax_data_df  =pd.read_csv("../Data/Clean/vax_data.csv")


vaers_ID_df = sample_event_data_df['VAERS_ID']

# Use VAERS_ID to filter and generate sample data

sample_sympt_data_df = pd.merge(vaers_ID_df, sympt_data_df, on='VAERS_ID', how='left')
sample_vax_data_df = pd.merge(vaers_ID_df, vax_data_df, on='VAERS_ID', how='left')


# Check shape of sample dataframes
print(sample_event_data_df.shape)
print(sample_sympt_data_df.shape)
print(sample_vax_data_df.shape)

# Save to csv
sample_sympt_data_df.to_csv('../Data/Clean/sample_sympt_data.csv', index=False)
sample_vax_data_df.to_csv('../Data/Clean/sample_vax_data.csv', index=False)
