==> Notebooks/3_SQL_Connector.ipynb <==
!pip install python-dotenv

!pip install --upgrade sqlalchemy

!pip install pymysql

import pandas as pd
import numpy as np
import pymysql
from sqlalchemy import create_engine


password = "Traynor_35"
bd = "covid_vax_events"
connection_string = "mysql+pymysql://root:" + password + "@localhost/"+bd
engine = create_engine(connection_string)
engine

from sqlalchemy import text
from dotenv import load_dotenv
import os

load_dotenv()

DB_NAME = os.getenv('DB_NAME')
DB_PASSWORD = os.getenv('DB_PASSWORD')

def query_to_df(sql):
    connection_string = "mysql+pymysql://root:" + DB_PASSWORD + "@localhost/" + DB_NAME
    engine = create_engine(connection_string)
    with engine.connect() as connection:
        query = text(sql)
        result = connection.execute(query)
        return pd.DataFrame(result.all())

sql_query = """
SELECT * FROM covid_vax_events.sample_vax_data;
"""
df=query_to_df(sql_query)

df

#save to csv
df.to_csv('../Data/Clean/sample_event_data.csv', index=False)
