==> Notebooks/4_Process_SQL_Data.ipynb <==
#Libraries
import pandas as pd
from functions import query_to_df

# Set Pandas display options to show all rows and columns
# pd.set_option("display.max_rows", None)  # None means no limit
# pd.set_option("display.max_columns", None)  # None means no limit

sql = """
SELECT 
    e.*,
    v.VAX_LOT, v.VAX_MANU, v.VAX_DOSE_SERIES,
    s.SYMPTOM1, s.SYMPTOM2, s.SYMPTOM3, s.SYMPTOM4, s.SYMPTOM5
FROM event_data AS e
LEFT JOIN sample_vax_data AS v
ON e.VAERS_ID = v.VAERS_ID
LEFT JOIN sample_sympt_data AS s
ON e.VAERS_ID = s.VAERS_ID
"""
df = query_to_df(sql)
df.head()

# Drop rows where 'VAX_LOT' or 'VAX_MANU' have null values
df = df.dropna(subset=['VAX_MANU'])

df

# Replace empty symptom fields with "None"
symptom_columns = ['SYMPTOM1', 'SYMPTOM2', 'SYMPTOM3', 'SYMPTOM4', 'SYMPTOM5']
df[symptom_columns] = df[symptom_columns].replace('', 'None')

# Combine SYMPTOM1 to SYMPTOM5 into a single column
df['ALL_SYMPTOMS'] = df[symptom_columns].apply(lambda x: ','.join(x.dropna().astype(str)), axis=1)


# Apply One-Hot Encoding on the combined symptoms
symptom_dummies = df['ALL_SYMPTOMS'].str.get_dummies(sep=',')

# Merge the one-hot encoded symptoms back into the original dataframe
df = pd.concat([df, symptom_dummies], axis=1)

# Display all column names
list(df.columns)

df

# 2035 cols 
# # Select the symptom columns
sympt_cols_df = df.iloc[:, -2035:]
# Calculate the occurrences of each symptom
symptom_counts = sympt_cols_df.sum(axis=0)

symptom_counts

# Create a DataFrame to display the column names and their counts
symptom_table = pd.DataFrame({
    "Symptom": symptom_counts.index,
    "Occurrences": symptom_counts.values
})
# Sort the table by occurrences in descending order
symptom_table_sorted = symptom_table.sort_values(by="Occurrences", ascending=False)
# Display the sorted table
display(symptom_table_sorted)

#Filter symptoms with more than 200 occurrences
symptoms_to_keep = symptom_table_sorted[symptom_table_sorted['Occurrences'] > 200]['Symptom'].tolist()
len(symptoms_to_keep)

# Get the original columns
original_cols = df.columns[:14].tolist()
original_cols

columns_to_keep = original_cols + symptoms_to_keep

df_processed = df[columns_to_keep]

df_processed

# Original columns without VAERS_ID
original_cols_without_id = [col for col in original_cols if col != 'VAERS_ID']
original_cols_without_id

# Define the aggregation dictionary
agg_dict = {col: "first" for col in original_cols_without_id}  # Max for all symptom columns
agg_dict.update({col: "max" for col in symptoms_to_keep})  # First for all other columns
# Group by ‘vars ID’ and apply the aggregation
df_merged = df_processed.groupby("VAERS_ID").agg(agg_dict).reset_index()
# Display the result
df_merged

df_merged

# Save the merged DataFrame to a new CSV file
df_merged.to_csv("../Data/Clean/full_dataset.csv", index=False)
